+++ lib/kekkan/models.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2010-2012 Arxopia LLC.
+# Copyright (c) 2012 Arxopia LLC.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -25,9 +25,14 @@
 #OF THE POSSIBILITY OF SUCH DAMAGE.
 
 module Kekkan
-	module CLI
+	module Models
 	end
 end
 
-require 'kekkan/cli/application'
-require 'kekkan/cli/banner'
+require 'kekkan/models/version'
+require 'kekkan/models/entry'
+require 'kekkan/models/cvss'
+require 'kekkan/models/reference'
+require 'kekkan/models/vulnerablesoftwarelist'
+require 'kekkan/models/assessmentcheck'
+require 'kekkan/models/scanner'