+++ NEWS.markdown
@@ -1,6 +1,8 @@
 # News
 
-#0.0.1 (August 1, 2012)
+#0.0.1 (October 13, 2012)
+- Added schema and SAX XML parser for NVD CVE v2 XML format
+- SQLite is very slow on large XML files, I recommend using MySql
 
 #0.0.0 (July 09, 2012)
 - Initial gem push to save the gem name, unusable